lib: crc32: make the crc_table variable non-const
When compiling with LTO, the compiler fails with an error saying that
`crc_table` causes a section type conflict with `efi_var_buf`.

This is because both are declared to be in the same section (via macro
`__efi_runtime_data`), but one is const while the other is not.

Make this variable non-const in order to fix this.

Signed-off-by: Marek Behún <marek.behun@nic.cz>